+++ common/index.js
@@ -1,0 +1,5 @@
+function returnDefaultValue(value, defaultValue) {
+  return value !== undefined ? value : defaultValue;
+}
+
+module.exports = { returnDefaultValue };